@@ -38,25 +38,31 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <spring.boot.version>2.7.10</spring.boot.version>
-        <mybatisplus.boot.version>3.5.3.1</mybatisplus.boot.version>
-        <mybatisplus.extension.version>3.5.3.1</mybatisplus.extension.version>
+        <java.version>21</java.version>
+        <spring.boot.version>3.5.4</spring.boot.version>
+        <mybatisplus.boot.version>3.5.12</mybatisplus.boot.version>
+        <mysql.version>9.3.0</mysql.version>
+        <commons.lang3.version>3.18.0</commons.lang3.version>
+        <gson.version>2.13.1</gson.version>
+        <lombok.version>1.18.38</lombok.version>
+        <logstash.version>7.3</logstash.version>
+        <guava.version>33.4.8-jre</guava.version>
+        <junit.version>4.13.2</junit.version>
+        <httpclient.version>5.5</httpclient.version>
+
         <jedis.version>3.8.0</jedis.version>
         <jaxb.core.version>4.0.2</jaxb.core.version>
         <jaxb.api.version>2.3.1</jaxb.api.version>
         <jaxb.impl.version>4.0.2</jaxb.impl.version>
         <activation.version>1.1.1</activation.version>
-        <mysql.version>8.0.32</mysql.version>
         <oracle.version>11.2.0.3</oracle.version>
         <quartz.version>2.3.2</quartz.version>
         <commons.lang.version>2.6</commons.lang.version>
-        <commons.lang3.version>3.12.0</commons.lang3.version>
         <commons.fileupload.version>1.5</commons.fileupload.version>
         <commons.io.version>1.3.2</commons.io.version>
         <commons.codec.version>1.15</commons.codec.version>
         <commons.configuration.version>1.10</commons.configuration.version>
-        <shiro.version>1.11.0</shiro.version>
+        <shiro.version>2.0.5</shiro.version>
         <jwt.version>0.9.1</jwt.version>
         <kaptcha.version>0.0.9</kaptcha.version>
 
@@ -66,22 +72,16 @@
 
         <swagger.version>3.0.0</swagger.version>
         <joda.time.version>2.12.5</joda.time.version>
-        <gson.version>2.10.1</gson.version>
         <velocity.version>1.7</velocity.version>
         <velocity.tools.version>2.0</velocity.tools.version>
         <easypoi.version>4.4.0</easypoi.version>
-        <lombok.version>1.18.26</lombok.version>
         <activiti.version>6.0.0</activiti.version>
         <batik.version>1.16</batik.version>
-        <logstash.version>7.3</logstash.version>
         <javamail.version>1.6.2</javamail.version>
         <actable.version>1.5.0.RELEASE</actable.version>
-        <guava.version>20.0</guava.version>
         <zxing.version>3.5.1</zxing.version>
-        <hibernate.validator.version>8.0.0.Final</hibernate.validator.version>
+        <validator.version>3.1.1</validator.version>
         <junit.version>4.13.2</junit.version>
-        <httpclient.version>4.5.14</httpclient.version>
-        <httpmime.version>4.5.14</httpmime.version>
         <spring.redis.version>1.4.7.RELEASE</spring.redis.version>
 
 
@@ -182,9 +182,11 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-compiler-plugin</artifactId>
-                    <version>3.8.1</version>
+                    <version>3.14.0</version>
                     <configuration>
-                        <compilerArgument>-Xlint:unchecked</compilerArgument>
+                        <source>21</source>
+                        <target>21</target>
+                        <encoding>UTF-8</encoding>
                     </configuration>
                 </plugin>
                 <!-- 跳过单元测试 -->
