@@ -11,7 +11,7 @@
  */
 package com.tongyi.common.validator.group;
 
-import javax.validation.GroupSequence;
+import jakarta.validation.GroupSequence;
 
 /**
  * 定义校验顺序，如果AddGroup组失败，则UpdateGroup组不会再校验
