@@ -6,8 +6,8 @@ import freemarker.template.TemplateNotFoundException;
 import org.springframework.beans.BeansException;
 import org.springframework.util.StringUtils;
 import org.springframework.web.servlet.view.AbstractTemplateView;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.util.Locale;
 import java.util.Map;
 
